requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-genai
fpdf2
python-multipart
